processor/requirements.txt
Pillow>=10.0.0
pytesseract>=0.3.10
PyPDF2>=3.0.0
pdf2image>=1.16.0


